+++ client/app/src/main/java/com/microsoft/research/karya/data/repo/MicroTaskRepository.kt
@@ -1,6 +1,6 @@
-package com.microsoft.research.karya.repo
+package com.microsoft.research.karya.data.repo
 
-import com.microsoft.research.karya.networking.MicroTaskAPI
+import com.microsoft.research.karya.data.service.MicroTaskAPI
 import kotlinx.coroutines.flow.flow
 import okhttp3.MultipartBody
