+++ client/app/src/main/java/com/microsoft/research/karya/data/service/MicroTaskAPI.kt
@@ -1,10 +1,15 @@
-package com.microsoft.research.karya.networking
+package com.microsoft.research.karya.data.service
 
-import com.microsoft.research.karya.database.models.KaryaFileRecord
+import com.microsoft.research.karya.data.model.karya.KaryaFileRecord
 import okhttp3.MultipartBody
 import okhttp3.ResponseBody
 import retrofit2.Response
-import retrofit2.http.*
+import retrofit2.http.GET
+import retrofit2.http.Header
+import retrofit2.http.Multipart
+import retrofit2.http.POST
+import retrofit2.http.Part
+import retrofit2.http.Path
 
 interface MicroTaskAPI {
